+++ NEWS.md
@@ -2,6 +2,8 @@
 
 Minor changes 
 
+* `report` now supports variables of class `htest` for the Friedman test.
+
 * `report` now supports variables of class `Date`, treating them like factors.
 
 * `report` now supports objects of class `estimate_contrasts`, from easystats'